refactor bindings, remove dead code
