+++ crates/bindings/src/lib.rs
@@ -1,3 +1,7 @@
+mod keys;
+
+pub use keys::*;
+
 use controller::ControllerButton;
 use directories::BaseDirs;
 use io::{AlphabeticKey, ArrowKey, Key, ModifierKey, MouseKey, NumericKey, SystemKey};
@@ -5,8 +9,6 @@ use io::{AlphabeticKey, ArrowKey, Key, ModifierKey, MouseKey, NumericKey, System
 use std::{collections::HashMap, sync::LazyLock};
 
 use serde::{Deserialize, Serialize};
-
-pub static JOYSTICK_KEYS: LazyLock<JoyStickKeys> = LazyLock::new(JoyStickKeys::default);
 
 #[rustfmt::skip]
 pub static CONTROLLER_KEY_MAP: LazyLock<HashMap<ControllerButton, Vec<Key>>> = LazyLock::new(|| {
@@ -147,49 +149,5 @@ mod bindings_map_format {
     }
 
     Ok(result)
-  }
-}
-
-#[derive(Debug, Clone)]
-pub struct JoyStickKeys {
-  forward: Vec<Key>,
-  backward: Vec<Key>,
-  port: Vec<Key>,
-  starboard: Vec<Key>,
-}
-
-impl Default for JoyStickKeys {
-  fn default() -> Self {
-    Self {
-      forward: CONTROLLER_KEY_MAP.get(&ControllerButton::Forward).cloned().unwrap_or_default(),
-      backward: CONTROLLER_KEY_MAP.get(&ControllerButton::Backward).cloned().unwrap_or_default(),
-      port: CONTROLLER_KEY_MAP.get(&ControllerButton::Port).cloned().unwrap_or_default(),
-      starboard: CONTROLLER_KEY_MAP.get(&ControllerButton::Starboard).cloned().unwrap_or_default(),
-    }
-  }
-}
-
-impl JoyStickKeys {
-  pub fn key_is_joystick_key(&self, key: Key) -> bool {
-    self.forward.contains(&key)
-      || self.backward.contains(&key)
-      || self.port.contains(&key)
-      || self.starboard.contains(&key)
-  }
-
-  pub fn forward(&self) -> &[Key] {
-    &self.forward
-  }
-
-  pub fn backward(&self) -> &[Key] {
-    &self.backward
-  }
-
-  pub fn port(&self) -> &[Key] {
-    &self.port
-  }
-
-  pub fn starboard(&self) -> &[Key] {
-    &self.starboard
   }
 }
