+++ crates/io/src/lib.rs
@@ -1,7 +1,3 @@
-mod event;
 mod key;
-mod movement;
 
-pub use event::*;
 pub use key::*;
-pub use movement::*;
