--- crates/io/src/movement/error.rs
@@ -1,4 +1,0 @@
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub enum MovementError {}
-
-impl std::error::Error for MovementError {}
