--- crates/io/src/movement/event.rs
@@ -1,6 +1,0 @@
-use crate::Movement;
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub struct MovementEvent {
-  movement: Movement,
-}
