+++ crates/platform/src/linux/joystick/axis.rs
@@ -27,3 +27,5 @@ pub fn try_from_relative_axis_code_for_axis(code: RelativeAxisCode) -> Result<Ax
   };
   Ok(axis)
 }
+
+
